educational_resources_technology/requirements.txt
Django
Pillow
django-environ
django-admin-extra-buttons